% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{abc}
\alias{abc}
\title{the ABC procedure for model selection}
\usage{
abc(y, g, x = 0)
}
\arguments{
\item{y}{the raw posttest scores of a continuous outcome variable.}

\item{g}{the categorical variable that denotes the group membership.}

\item{x}{(optional) the raw pretest scores of a continuous outcome variable.}
}
\value{
AIC results of different models. The model with smallest AIC is preferred.
}
\description{
the AIC comparison with Modified Box-Cox Transformation (ABC)
is a diagnostic procedure to help select among various additive
and multiplicative models
}
\details{
When only `y` and `g` are specified, the ABC procedure compares
LANOVA and ANOVA models. When `x` is also specified, the ABC procedure
compares LANCOVA, ANCOVA, ANCOHET, and ANCOVA with log-transformed y.
}
\examples{
data("schoene")
attach(schoene)
abc(post_HRT,group,pre_HRT)
abc(post_HRT,group)
}
