% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamme-package.R
\docType{package}
\name{lamme}
\alias{lamme}
\alias{lamme-package}
\title{lamme}
\description{
Log-Analytic Methods for Multiplicative Effects
}
\details{
The \code{lamme} package is designed to test and estimate multiplicative effects
via log-analytic methods.
}
\section{Usage}{

To access this package's tutorial, type the following line into the console:

\code{vignette("lamme-vignette")}
}

