% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanova.R
\name{lanova}
\alias{lanova}
\title{Logged ANOVA}
\usage{
lanova(y, g, plot = F)
}
\arguments{
\item{y}{the raw scores of a continuous outcome variable.}

\item{g}{a categorical variable that denotes the group membership.}

\item{plot}{a TRUE/FALSE variable that denotes if diagnostic plots are desired. (default=F)}
}
\value{
An summary object of the LANOVA results.
In residuals, the summary statistics are of sample multiplicative errors.
In the coefficients table, the estimate of the intercept is the default group (control group) geometric mean estimate.
Other coefficient estimates are effect size measure zeta's estiamtes. The standard error is on the logged scale.
The confidence intervals are of 95% for the control group geometric mean and for the zeta estimates,
respectively, of the intercept and otherwise.
The residual standard error is that of the logged scale residuals. Both R-squared and Ajudsted R-squared are computed
on the logged model. If `plot==TRUE`, diagnostic plots are provided.
}
\description{
Mathematically, LANOVA is the ANOVA form of
a log-log model where the dependent variable is log-transformed.
LANOVA can test and estimate multiplicative effects.
}
\examples{
# generate data
y1=rnorm(1000,5,1)+rnorm(1000)
y2=rnorm(1000,5.5,1)+rnorm(1000)
y3=rnorm(1000,6,1)+rnorm(1000)
y1=exp(y1)
y2=exp(y2)
y3=exp(y3)
dep=c(y1,y2,y3)
tc=rep(c(0,1,2),each=1000)
# applying lanova with the generated data
lanova(dep,tc)
}
