% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr.lancova.R
\name{pwr.lancova}
\alias{pwr.lancova}
\title{Power Calculation for LANCOVA}
\usage{
pwr.lancova(k, n, r_sqrd, rho_sqrd, alpha = 0.05)
}
\arguments{
\item{k}{the number of groups.}

\item{n}{the number of observations per group.}

\item{r_sqrd}{the expected explained variance (on the logged scale)}

\item{rho_sqrd}{the pretest-posttest correlation}

\item{alpha}{the significance level (default=.05)}
}
\value{
\item{power}{the statistical power of test}
}
\description{
Compute the statistical power of the LANCOVA test.
}
\examples{
pwr.lancova(3,40,.1,.4,.05)
}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). Hillsdale,NJ: Lawrence Erlbaum.
}
