% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr.lanova.R
\name{pwr.lanova}
\alias{pwr.lanova}
\title{Power Calculation for LANOVA}
\usage{
pwr.lanova(k, n, r_sqrd, alpha = 0.05)
}
\arguments{
\item{k}{the number of groups.}

\item{n}{the number of observations per group.}

\item{r_sqrd}{the expected explained variance (on the logged scale)}

\item{alpha}{the significance level (default=.05)}
}
\value{
\item{power}{the statistical power of test}
}
\description{
Compute the statistical power of the LANOVA test.
}
\examples{
pwr.lanova(3,40,.4,.05)
}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). Hillsdale,NJ: Lawrence Erlbaum.
}
