% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.lancova.R
\name{ss.lancova}
\alias{ss.lancova}
\title{Sample Size Planning for LANCOVA}
\usage{
ss.lancova(k, rho_sqrd, r_sqrd, power = 0.8, alpha = 0.05)
}
\arguments{
\item{k}{the number of groups.}

\item{rho_sqrd}{the pretest-posttest correlation}

\item{r_sqrd}{the expected explained variance by the model (on the logged scale)}

\item{power}{the desired statistical power (default=.8)}

\item{alpha}{the significance level (default=.05)}
}
\value{
\item{n}{the per-group sample size requirement}
}
\description{
Compute the required per-group sample size for the LANCOVA test.
}
\examples{
ss.lancova(3,.5,.01,.14,.05)
}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). Hillsdale,NJ: Lawrence Erlbaum.
}
