% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.lanova.R
\name{ss.lanova}
\alias{ss.lanova}
\title{Sample Size Planning for LANOVA}
\usage{
ss.lanova(k, r_sqrd, power = 0.8, alpha = 0.05)
}
\arguments{
\item{k}{the number of groups.}

\item{r_sqrd}{the expected explained variance (on the logged scale)}

\item{power}{the desired statistical power (default=.8)}

\item{alpha}{the significance level (default=.05)}
}
\value{
\item{n}{the per-group sample size requirement}
}
\description{
Compute the required per-group sample size for the LANOVA test.
}
\examples{
ss.lanova(3,.01,.14,.05)
}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). Hillsdale,NJ: Lawrence Erlbaum.
}
